[refactor] Change function arg type

+++ srcs/ft_striteri.c
@@ -6,15 +6,15 @@
 /*   By: ttomori <ttomori@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/09 01:25:55 by ttomori           #+#    #+#             */
-/*   Updated: 2022/01/20 00:06:14 by ttomori          ###   ########.fr       */
+/*   Updated: 2022/01/20 00:14:21 by ttomori          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_striteri(char *s, void (*f)(unsigned int, char *))
+void	ft_striteri(char *s, void (*f)(size_t, char *))
 {
-	unsigned int	i;
+	size_t	i;
 
 	if (f == NULL || s == NULL)
 		return ;
